AE: refactor IAudioCallback
