+++ xbmc/cores/AudioEngine/Interfaces/AEStream.h
@@ -20,7 +20,7 @@
  */
 
 #include "cores/AudioEngine/Utils/AEAudioFormat.h"
-#include "cores/IAudioCallback.h"
+#include "cores/AudioEngine/Interfaces/IAudioCallback.h"
 #include <stdint.h>
 
 extern "C" {
